fix: move domains submodule to separate file
